fix(entity): Wolf 愤怒字段 wire 类型 i32→i64(Long) 修复真客户端 field21 类型校验崩
vanilla 1.21.11 Wolf.DATA_ANGER_END_TIME 为 Long(serializerId=2 VAR_LONG,默认 -1L),
项目 DATA_ANGER_TIME_PARAM 误用 i32(serializerId=1 INT,默认 0),真客户端 set_entity_data
field21 类型校验崩(disconnect-2026-08-04: old=-1(Long) new=0(Integer))。

DATA_ANGER_TIME_PARAM 改 i64 对齐 wire Long 类型;业务层仍以「剩余 ticks」i32 语义运转
(getAngerTime/setAngerTime 做 i32↔i64 互转,值域 0..MAX_ANGER_TIME 作为 Long 合法,
客户端按 Long 读再 >0 判 angry 自洽),IAngerable 接口与 TameableEntity 基类零改动。
客户端 ClientEntity 改读 i64;WolfEntityTest 改读 i64;黄金表测试补 field21 必为
i64/Long(variant index 2)断言防回归(原仅校验 id 区间漏网致此 bug)。

+++ src/client/world/entity/ClientEntity.cpp
@@ -319,9 +319,10 @@ void ClientEntity::syncMetadataFromDataManager()
             setWolfIsInterested(*interested);
         }
         // 愤怒状态（尾巴抬起/停止摆动 + angry 纹理变体，通过 WolfEntity::DATA_ANGER_TIME_PARAM 同步）
+        // 该参数对齐 vanilla Wolf.DATA_ANGER_END_TIME，wire 类型为 Long(i64)，故读 i64。
         if (m_dataManager.hasParam(::mc::WolfEntity::getAngerTimeParamId())) {
             if (const auto* value = m_dataManager.getRaw(::mc::WolfEntity::getAngerTimeParamId()); value != nullptr) {
-                const i32 angerTime = value->get<i32>();
+                const i64 angerTime = value->get<i64>();
                 setWolfIsAngry(angerTime > 0);
             }
         }
